+++ modules/tasks.js
@@ -1,0 +1,9 @@
+class Tasks {
+  constructor(index, description, completed) {
+    this.index = index;
+    this.description = description;
+    this.completed = completed;
+  }
+}
+
+export default Tasks;
