+++ src/index.html
@@ -27,7 +27,7 @@
             <form class="task-input-form">
                 <input type="text" name="input-task" placeholder="Add to your list..." spellcheck="false"
                     id="input-item">
-                <div class="add-item" type="submit">
+                <div class="add-item">
                     <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor"
                         class="bi bi-arrow-return-left" viewBox="0 0 16 16">
                         <path fill-rule="evenodd"
@@ -37,7 +37,7 @@
             </form>
         </li>
         <li class="clear-completed">
-            <p><button class="btn-clear">Clear all completed</button></p>
+            <button class="btn-clear">Clear all completed</button>
         </li>
     </ul>
 </body>
